+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Gandi V5 API Gem Changelog
 
+## Version 0.5.0
+
+*
+
 ## Version 0.4.0
 
 * Fix exception when delete returns no content-type